Added API for rejecting dm invites.
