+++ src/main.rs
@@ -31,9 +31,11 @@ async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
             .service(api::server::join_server)                      // change token !!
             .service(api::server::get_server_users)                 
             .service(api::server::get_server_info)
-            .service(api::server::send_dm_invite)
-            .service(api::server::accept_dm_invite)
-            .service(api::server::fetch_pending_dm_invites)
+            
+            .service(api::invites::send_dm_invite)
+            .service(api::invites::accept_dm_invite)
+            .service(api::invites::reject_dm_invite)
+            .service(api::invites::fetch_pending_dm_invites)
 
 
             .service(api::channel::get_channels)
